+++ NEWS.md
@@ -10,6 +10,7 @@
 * Add silica to PlotBasinOutput() and PlotBasinSummary()
 * Add observed snow water equivalent (rswe) to PlotBasinOutput()
 * Updated the demo model files to e.g. add POURX and POURY to GeoData.txt based on the subbasin centroids
+* Ensure that input is converted to dataframe for AnnualRegime()
 
 ## HYPEtools 1.6.6 (2025-08-19)
 *Enhancements*